+++ CHANGES.md
@@ -13,6 +13,7 @@ Release Notes.
 * Fix ClassCastException when SQLServer inserts data 
 * [Chore] Exclude org.checkerframework:checker-qual and com.google.j2objc:j2objc-annotations
 * [Chore] Exclude proto files in the generated jar
+* Support Grizzly Trace
 
 #### Documentation